chapter 9

+++ src/ray.rs
@@ -20,7 +20,7 @@ impl Ray {
         self.b
     }
 
-    pub fn point_at_parameter(&self, t: f64) -> Vec3 {
+    pub fn point_at_parameter(&self, t: f32) -> Vec3 {
         self.a + self.b * t
     }
 }
